Completely rewrite Executors, so parent table structure can be supported
This rewrite is still WIP.
Basic programming principles like the 'single responsibility' principle have to be reapplied this code.
Existing validators have to be reviewed and possibly rewritten/updated to account for parent tables.
Readme file has to be updated, as well as the javadoc on the mojo.

Unit tests for the new validators have to be added.
Unit tests for the util methods have to be added.